automl_agents_2/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
python-dotenv>=1.0.0
